+++ CHANGES.md
@@ -1,5 +1,6 @@
 # Unreleased
 - [NEW] Handle HTTP status code `429 Too Many Requests` with blocking backoff and retries.
+- [NEW] Added `DesignDocumentManager.list()` to return all design documents defined in a database.
 
 # 2.4.3 (2016-05-05)
 - [IMPROVED] Reduced the length of the User-Agent header string.